+++ DataServices/Interfaces/IFileDS.cs
@@ -2,9 +2,9 @@ using DriveAPI.Models;
 
 namespace DriveAPI.DataServices{
     public interface IFileDS{
-        public Filew GetFile(int id);
+        public Filew GetFile(long id);
         public Filew PostFile(Filew file);
-        public Filew PutFile(int id, Filew file);
-        public int DeleteFile(int id);
+        public Filew PutFile(long id, Filew file);
+        public long DeleteFile(long id);
     }
 }
